Modify Bury and Claim
The bury and claim functions are modified in order to move away from burying prl in actual ethereum addresses to burying them directly in the smart contract. This is achieved by burying the prl with a hash (which acts as the public key) and requiring the website owner to call the claim function with the key of that hash (which acts as the public key).  Other minor modifications as actually unburying the hash when calling the claim function, making the _transfer function return a bool on success have also been made.